Moved tests from issues to methods_and_attributes
