+++ tests/test_methods_and_attributes.py
@@ -255,3 +255,27 @@ def test_noconvert_args(msg):
 
         Invoked with: 4
     """
+
+
+def test_bad_arg_default(msg):
+    from pybind11_tests import debug_enabled, bad_arg_def_named, bad_arg_def_unnamed
+
+    with pytest.raises(RuntimeError) as excinfo:
+        bad_arg_def_named()
+    assert msg(excinfo.value) == (
+        "arg(): could not convert default argument 'a: NotRegistered' in function 'should_fail' "
+        "into a Python object (type not registered yet?)"
+        if debug_enabled else
+        "arg(): could not convert default argument into a Python object (type not registered "
+        "yet?). Compile in debug mode for more information."
+    )
+
+    with pytest.raises(RuntimeError) as excinfo:
+        bad_arg_def_unnamed()
+    assert msg(excinfo.value) == (
+        "arg(): could not convert default argument 'NotRegistered' in function 'should_fail' "
+        "into a Python object (type not registered yet?)"
+        if debug_enabled else
+        "arg(): could not convert default argument into a Python object (type not registered "
+        "yet?). Compile in debug mode for more information."
+    )
